Path Implem
In the previous version of YOCaml, abstraction on paths was a problem.
However, rather than abstracting paths by interpreting effects (which
requires the effect manager to be functorised, which is ... very
cumbersome), this commit proposes a standard representation of a file
path. The abstraction comes at the level of consumption, transforming an
expression of type `Path.t` into something understandable by a Runtime.
The implementation of paths is, for the moment, relatively naive for
crystallising its API.

Co-authored-by: xvw <xaviervdw@gmail.com>