Add Test effect for acting on Time in test ctx
In Mini_yocaml (see [1]) time is incremented with each file write
because the test framework does not add an effect to control time. This
makes the status of artifact modification dates implicit and complicated
to track. So rather than using an ad-hoc strategy, we add an effect,
only in the test scope, to allow the test-level to control the time.

[1] https://github.com/xvw/mini_yocaml/

Co-authored-by: xvw <xaviervdw@gmail.com>