A first attempt of dynamic dependencies handling
In early versions of YOCaml (and Mini_YOCaml), the management of dynamic
dependencies was delegated to the user, making interaction between
static and dynamic dependencies complicated. To interweave static and
dynamic dependencies in an ergonomic way, the task to be executed is
given the option of _returning_ a list of dynamic dependencies and
relying on the cache to decide whether or not to rebuild a file.

In some cases, if the cache doesn't exist, a reperformance of effect is
necessary, however, we assume that the cache is correctly maintained,
because we're... optimistic.

Although it wasn't present in Mini_YOCaml, the global algorithm for
maximizing minimality (lol) comes from @xvw.

Co-authored-by: xvw <xaviervdw@gmail.com>