Archetype: Datetime
Archetypes describe generic models that can be used if you don't want to
describe your models manually, and should make it easier to describe
blogs without declaring metadata. It replaces the `Metadata` module
historically present in Yocaml1 and can be used either as an example (of
how to describe data) or directly when you're a bit lazy. However,
modules such as `Datetime` can also be used in custom templates.

The code is largely ported from YOCaml's historical Date code, but also
from the LADMM-SA application code.

Co-authored-by: xvw <xaviervdw@gmail.com>
Co-authored-by: grrm <grimfw@gmail.com>