Update mtime (check recursion)
As @grrm mentioned, using YOCaml's dynamic dependencies as they are
makes index processing complicated. In an index, we don't rely on a set
of "known" files or "described by reading a file", but on the presence
of files in a directory, for example. There are several solutions to
this problem, but a naive approach would be to consider that a
directory, as a dependency, will be based on the greatest modification
date of its children. In this way, we could simply add the path
"./articles" as a dependency and adding a file to this directory would
change its modification date.

Co-authored-by: xvw <xaviervdw@gmail.com>
Co-authored-by: grrm <grimfw@gmail.com>