Store Build-stamp in Cache
Before this commit, if a file is modified without providing any concrete
modification (ie: `update + save + undo + save`), the hash in the cache
is continuously calculated and read (until a concrete modification
affects the construction). With this patch, the entries in the cache are
stamped so that, once the hashes have been compared, the construction is
validated for future calls.