raise `exit` for keeping server alive
When you're writing an article, a metadata error can lead to a server
shutdown, which can be very annoying. So I suggest throwing an exception
which, in the event of a build, will return an error, but in the event
of a live-server, will be caught by the exception handler. The
modification has no impact on the unix server or the EIO server.