Return true for all queries about the type of an element as a directory
If we assume that we're handling a well-formed Git repository, there's no need
to check whether an element is actually a directory or not. We could do a real
check (and see that the element is a tree) but it would be a little slower.

The problem with the current code is that an existing folder is then considered
to be a file: and YOcaml fails.